requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
amqp==5.0.2
    # via kombu
anyascii==0.1.7
    # via wagtail
asgiref==3.3.4
    # via django
beautifulsoup4==4.8.2
    # via wagtail
billiard==3.6.3.0
    # via celery
celery==5.0.2
    # via -r requirements.in
certifi==2020.6.20
    # via
    #   requests
    #   sentry-sdk
chardet==3.0.4
    # via requests
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-repl
click-didyoumean==0.0.3
    # via celery
click-repl==0.1.6
    # via celery
django==3.2.4
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-filter
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   wagtail
django-ckeditor==6.0.0
    # via -r requirements.in
django-cors-headers==3.5.0
    # via -r requirements.in
django-debug-toolbar==3.2.1
    # via -r requirements.in
django-environ==0.4.5
    # via -r requirements.in
django-filter==2.4.0
    # via
    #   -r requirements.in
    #   wagtail
django-js-asset==1.2.2
    # via django-ckeditor
django-modelcluster==5.1
    # via
    #   -r requirements.in
    #   wagtail
django-taggit==1.4.0
    # via
    #   -r requirements.in
    #   wagtail
django-treebeard==4.3.1
    # via wagtail
djangorestframework==3.12.1
    # via
    #   -r requirements.in
    #   wagtail
djangorestframework-serializer-extensions==2.0.1
    # via -r requirements.in
draftjs-exporter==2.1.7
    # via wagtail
et-xmlfile==1.0.1
    # via openpyxl
hashids==1.3.1
    # via djangorestframework-serializer-extensions
html5lib==1.1
    # via wagtail
idna==2.10
    # via requests
importlib-metadata==2.0.0
    # via kombu
inflect==5.3.0
    # via -r requirements.in
jdcal==1.4.1
    # via openpyxl
kombu==5.0.2
    # via celery
l18n==2020.6.1
    # via wagtail
numpy==1.19.4
    # via pandas
openpyxl==3.0.5
    # via tablib
pandas==1.1.5
    # via -r requirements.in
pillow==8.2.0
    # via wagtail
prompt-toolkit==3.0.8
    # via click-repl
psycopg2==2.8.6
    # via -r requirements.in
python-dateutil==2.8.1
    # via
    #   -r requirements.in
    #   pandas
pytz==2020.1
    # via
    #   celery
    #   django
    #   django-modelcluster
    #   l18n
    #   pandas
requests==2.25.1
    # via
    #   -r requirements.in
    #   wagtail
sentry-sdk==1.3.0
    # via -r requirements.in
six==1.15.0
    # via
    #   click-repl
    #   html5lib
    #   l18n
    #   python-dateutil
soupsieve==2.0.1
    # via beautifulsoup4
sqlparse==0.4.1
    # via
    #   django
    #   django-debug-toolbar
tablib[xls,xlsx]==2.0.0
    # via wagtail
telepath==0.2
    # via wagtail
typing-extensions==3.10.0.0
    # via asgiref
urllib3==1.26.5
    # via
    #   requests
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
wagtail==2.13.2
    # via -r requirements.in
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via html5lib
willow==1.4
    # via wagtail
xlrd==1.2.0
    # via tablib
xlsxwriter==1.3.7
    # via
    #   -r requirements.in
    #   wagtail
xlwt==1.3.0
    # via tablib
zipp==3.4.0
    # via importlib-metadata
